Regorganised file structure
