+++ src/Ogooreck.Sample.BusinessLogic.Tests/Aggregates/EventSourced/Pricing/IProductPriceCalculator.cs
@@ -1,0 +1,8 @@
+using Ogooreck.Sample.BusinessLogic.Aggregates.EventSourced.Products;
+
+namespace Ogooreck.Sample.BusinessLogic.Aggregates.EventSourced.Pricing;
+
+public interface IProductPriceCalculator
+{
+    IReadOnlyList<PricedProductItem> Calculate(params ProductItem[] productItems);
+}
