+++ src/Ogooreck.Sample.BusinessLogic.Tests/Aggregates/EventSourced/ShoppingCart.cs
@@ -1,0 +1,158 @@
+using Ogooreck.Sample.BusinessLogic.Aggregates.EventSourced.Core;
+using Ogooreck.Sample.BusinessLogic.Aggregates.EventSourced.Pricing;
+using Ogooreck.Sample.BusinessLogic.Aggregates.EventSourced.Products;
+
+namespace Ogooreck.Sample.BusinessLogic.Aggregates.EventSourced;
+
+public class ShoppingCart: Aggregate
+{
+    public Guid ClientId { get; private set; }
+
+    public ShoppingCartStatus Status { get; private set; }
+
+    public IList<PricedProductItem> ProductItems { get; private set; } = default!;
+
+    public decimal TotalPrice => ProductItems.Sum(pi => pi.TotalPrice);
+
+    public static ShoppingCart Open(
+        Guid cartId,
+        Guid clientId)
+    {
+        return new ShoppingCart(cartId, clientId);
+    }
+
+    private ShoppingCart(){}
+
+    private ShoppingCart(
+        Guid id,
+        Guid clientId)
+    {
+        var @event = ShoppingCartOpened.Create(
+            id,
+            clientId
+        );
+
+        Enqueue(@event);
+        Apply(@event);
+    }
+
+    public void Apply(ShoppingCartOpened @event)
+    {
+        Id = @event.CartId;
+        ClientId = @event.ClientId;
+        ProductItems = new List<PricedProductItem>();
+        Status = ShoppingCartStatus.Pending;
+    }
+
+    public void AddProduct(
+        IProductPriceCalculator productPriceCalculator,
+        ProductItem productItem)
+    {
+        if(Status != ShoppingCartStatus.Pending)
+            throw new InvalidOperationException($"Adding product for the cart in '{Status}' status is not allowed.");
+
+        var pricedProductItem = productPriceCalculator.Calculate(productItem).Single();
+
+        var @event = ProductAdded.Create(Id, pricedProductItem);
+
+        Enqueue(@event);
+        Apply(@event);
+    }
+
+    public void Apply(ProductAdded @event)
+    {
+        var newProductItem = @event.ProductItem;
+
+        var existingProductItem = FindProductItemMatchingWith(newProductItem);
+
+        if (existingProductItem is null)
+        {
+            ProductItems.Add(newProductItem);
+            return;
+        }
+
+        ProductItems.Replace(
+            existingProductItem,
+            existingProductItem.MergeWith(newProductItem)
+        );
+    }
+
+    public void RemoveProduct(
+        PricedProductItem productItemToBeRemoved)
+    {
+        if(Status != ShoppingCartStatus.Pending)
+            throw new InvalidOperationException($"Removing product from the cart in '{Status}' status is not allowed.");
+
+        var existingProductItem = FindProductItemMatchingWith(productItemToBeRemoved);
+
+        if (existingProductItem is null)
+            throw new InvalidOperationException($"Product with id `{productItemToBeRemoved.ProductId}` and price '{productItemToBeRemoved.UnitPrice}' was not found in cart.");
+
+        if(!existingProductItem.HasEnough(productItemToBeRemoved.Quantity))
+            throw new InvalidOperationException($"Cannot remove {productItemToBeRemoved.Quantity} items of Product with id `{productItemToBeRemoved.ProductId}` as there are only ${existingProductItem.Quantity} items in card");
+
+        var @event = ProductRemoved.Create(Id, productItemToBeRemoved);
+
+        Enqueue(@event);
+        Apply(@event);
+    }
+
+    public void Apply(ProductRemoved @event)
+    {
+        var productItemToBeRemoved = @event.ProductItem;
+
+        var existingProductItem = FindProductItemMatchingWith(@event.ProductItem);
+
+        if (existingProductItem == null)
+            return;
+
+        if (existingProductItem.HasTheSameQuantity(productItemToBeRemoved))
+        {
+            ProductItems.Remove(existingProductItem);
+            return;
+        }
+
+        ProductItems.Replace(
+            existingProductItem,
+            existingProductItem.Subtract(productItemToBeRemoved)
+        );
+    }
+
+    public void Confirm()
+    {
+        if(Status != ShoppingCartStatus.Pending)
+            throw new InvalidOperationException($"Confirming cart in '{Status}' status is not allowed.");
+
+        var @event = ShoppingCartConfirmed.Create(Id, DateTime.UtcNow);
+
+        Enqueue(@event);
+        Apply(@event);
+    }
+
+    public void Apply(ShoppingCartConfirmed @event)
+    {
+        Status = ShoppingCartStatus.Confirmed;
+    }
+
+    public void Cancel()
+    {
+        if(Status != ShoppingCartStatus.Pending)
+            throw new InvalidOperationException($"Canceling cart in '{Status}' status is not allowed.");
+
+        var @event = ShoppingCartCanceled.Create(Id, DateTime.UtcNow);
+
+        Enqueue(@event);
+        Apply(@event);
+    }
+
+    public void Apply(ShoppingCartCanceled @event)
+    {
+        Status = ShoppingCartStatus.Canceled;
+    }
+
+    private PricedProductItem? FindProductItemMatchingWith(PricedProductItem productItem)
+    {
+        return ProductItems
+            .SingleOrDefault(pi => pi.MatchesProductAndPrice(productItem));
+    }
+}
