+++ src/Ogooreck.Sample.BusinessLogic.Tests/Aggregates/StateBased/Pricing/IProductPriceCalculator.cs
@@ -1,0 +1,8 @@
+using Ogooreck.Sample.BusinessLogic.Aggregates.StateBased.Products;
+
+namespace Ogooreck.Sample.BusinessLogic.Aggregates.StateBased.Pricing;
+
+public interface IProductPriceCalculator
+{
+    IReadOnlyList<PricedProductItem> Calculate(params ProductItem[] productItems);
+}
