+++ src/Ogooreck.Sample.BusinessLogic.Tests/Aggregates/StateBased/ShoppingCartStatus.cs
@@ -1,0 +1,8 @@
+namespace Ogooreck.Sample.BusinessLogic.Aggregates.StateBased;
+
+public enum ShoppingCartStatus
+{
+    Pending = 1,
+    Confirmed = 2,
+    Canceled = 4
+}
