+++ src/Ogooreck.Sample.BusinessLogic.Tests/Functions/Incident.cs
@@ -1,0 +1,164 @@
+namespace Ogooreck.Sample.BusinessLogic.Functions;
+
+public record IncidentLogged(
+    Guid IncidentId,
+    Guid CustomerId,
+    Contact Contact,
+    string Description,
+    Guid LoggedBy,
+    DateTimeOffset LoggedAt
+);
+
+public record IncidentCategorised(
+    Guid IncidentId,
+    IncidentCategory Category,
+    Guid CategorisedBy,
+    DateTimeOffset CategorisedAt
+);
+
+public record IncidentPrioritised(
+    Guid IncidentId,
+    IncidentPriority Priority,
+    Guid PrioritisedBy,
+    DateTimeOffset PrioritisedAt
+);
+
+public record AgentAssignedToIncident(
+    Guid IncidentId,
+    Guid AgentId,
+    DateTimeOffset AssignedAt
+);
+
+public record AgentRespondedToIncident(
+    Guid IncidentId,
+    IncidentResponse.FromAgent Response,
+    DateTimeOffset RespondedAt
+);
+
+public record CustomerRespondedToIncident(
+    Guid IncidentId,
+    IncidentResponse.FromCustomer Response,
+    DateTimeOffset RespondedAt
+);
+
+public record IncidentResolved(
+    Guid IncidentId,
+    ResolutionType Resolution,
+    Guid ResolvedBy,
+    DateTimeOffset ResolvedAt
+);
+
+public record ResolutionAcknowledgedByCustomer(
+    Guid IncidentId,
+    Guid AcknowledgedBy,
+    DateTimeOffset AcknowledgedAt
+);
+
+public record IncidentClosed(
+    Guid IncidentId,
+    Guid ClosedBy,
+    DateTimeOffset ClosedAt
+);
+
+public enum IncidentStatus
+{
+    Pending = 1,
+    Resolved = 8,
+    ResolutionAcknowledgedByCustomer = 16,
+    Closed = 32
+}
+
+public record Incident(
+    Guid Id,
+    IncidentStatus Status = IncidentStatus.Pending,
+    bool HasOutstandingResponseToCustomer = false,
+    IncidentCategory? Category = null,
+    IncidentPriority? Priority = null,
+    Guid? AgentId = null,
+    int Version = 1
+)
+{
+    public static Incident Create(IncidentLogged logged) =>
+        new(logged.IncidentId);
+
+    public Incident Apply(IncidentCategorised categorised) =>
+        this with { Category = categorised.Category };
+
+    public Incident Apply(IncidentPrioritised prioritised) =>
+        this with { Priority = prioritised.Priority };
+
+    public Incident Apply(AgentRespondedToIncident agentResponded) =>
+        this with { HasOutstandingResponseToCustomer = false };
+
+    public Incident Apply(CustomerRespondedToIncident customerResponded) =>
+        this with { HasOutstandingResponseToCustomer = true };
+
+    public Incident Apply(IncidentResolved resolved) =>
+        this with { Status = IncidentStatus.Resolved };
+
+    public Incident Apply(ResolutionAcknowledgedByCustomer acknowledged) =>
+        this with { Status = IncidentStatus.ResolutionAcknowledgedByCustomer };
+
+    public Incident Apply(IncidentClosed closed) =>
+        this with { Status = IncidentStatus.Closed };
+}
+
+public enum IncidentCategory
+{
+    Software,
+    Hardware,
+    Network,
+    Database
+}
+
+public enum IncidentPriority
+{
+    Critical,
+    High,
+    Medium,
+    Low
+}
+
+public enum ResolutionType
+{
+    Temporary,
+    Permanent,
+    NotAnIncident
+}
+
+public enum ContactChannel
+{
+    Email,
+    Phone,
+    InPerson,
+    GeneratedBySystem
+}
+
+public record Contact(
+    ContactChannel ContactChannel,
+    string? FirstName = null,
+    string? LastName = null,
+    string? EmailAddress = null,
+    string? PhoneNumber = null
+);
+
+public abstract record IncidentResponse
+{
+    public record FromAgent(
+        Guid AgentId,
+        string Content,
+        bool VisibleToCustomer
+    ): IncidentResponse(Content);
+
+    public record FromCustomer(
+        Guid CustomerId,
+        string Content
+    ): IncidentResponse(Content);
+
+    public string Content { get; init; }
+
+    private IncidentResponse(string content)
+    {
+        Content = content;
+    }
+}
