--- src/Ogooreck.Sample.EventSourcing/Aggregates/EventSourced/Pricing/IProductPriceCalculator.cs
@@ -1,8 +1,0 @@
-using Ogooreck.Sample.EventSourcing.Aggregates.EventSourced.Products;
-
-namespace Ogooreck.Sample.EventSourcing.Aggregates.EventSourced.Pricing;
-
-public interface IProductPriceCalculator
-{
-    IReadOnlyList<PricedProductItem> Calculate(params ProductItem[] productItems);
-}
