--- src/Ogooreck.Sample.EventSourcing/Aggregates/EventSourced/ShoppingCartStatus.cs
@@ -1,8 +1,0 @@
-namespace Ogooreck.Sample.EventSourcing.Aggregates.EventSourced;
-
-public enum ShoppingCartStatus
-{
-    Pending = 1,
-    Confirmed = 2,
-    Canceled = 4
-}
