--- src/Ogooreck.Sample.EventSourcing/Aggregates/StateBased/Core/Aggregate.cs
@@ -1,12 +1,0 @@
-namespace Ogooreck.Sample.EventSourcing.Aggregates.StateBased.Core;
-
-public abstract class Aggregate: Aggregate<Guid>
-{
-}
-
-public abstract class Aggregate<T> where T : notnull
-{
-    public T Id { get; protected set; } = default!;
-
-    public int Version { get; protected set; }
-}
