--- src/Ogooreck.Sample.EventSourcing/Aggregates/StateBased/Pricing/IProductPriceCalculator.cs
@@ -1,8 +1,0 @@
-using Ogooreck.Sample.EventSourcing.Aggregates.StateBased.Products;
-
-namespace Ogooreck.Sample.EventSourcing.Aggregates.StateBased.Pricing;
-
-public interface IProductPriceCalculator
-{
-    IReadOnlyList<PricedProductItem> Calculate(params ProductItem[] productItems);
-}
