+++ src/Ogooreck.sln
@@ -29,7 +29,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "mdsource", "mdsource", "{2B
 		..\mdsource\README.source.md = ..\mdsource\README.source.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ogooreck.Sample.EventSourcing", "Ogooreck.Sample.EventSourcing\Ogooreck.Sample.EventSourcing.csproj", "{0079115B-9C5E-4C03-9215-0CA6B418340E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ogooreck.Sample.BusinessLogic.Tests", "Ogooreck.Sample.BusinessLogic.Tests\Ogooreck.Sample.BusinessLogic.Tests.csproj", "{90A3027A-C4BA-4141-967B-4FD10B9AADC3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,10 +53,10 @@ Global
 		{5D676ACB-C70F-447A-94AF-4547889CA100}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5D676ACB-C70F-447A-94AF-4547889CA100}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5D676ACB-C70F-447A-94AF-4547889CA100}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0079115B-9C5E-4C03-9215-0CA6B418340E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0079115B-9C5E-4C03-9215-0CA6B418340E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0079115B-9C5E-4C03-9215-0CA6B418340E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0079115B-9C5E-4C03-9215-0CA6B418340E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{90A3027A-C4BA-4141-967B-4FD10B9AADC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{90A3027A-C4BA-4141-967B-4FD10B9AADC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{90A3027A-C4BA-4141-967B-4FD10B9AADC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{90A3027A-C4BA-4141-967B-4FD10B9AADC3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{5D676ACB-C70F-447A-94AF-4547889CA100} = {F617812B-14C9-45AB-BDE1-5CDA13C339FA}
